Improve parsing of image content
I was tracking down a bug in our app where users would paste a url to an
image in a comment field and see a garbled summary of the contents
behind that url. We use metainspector in the backend to scrape urls, and
after digging around a bit I discovered that any url with content type
`image/*` would result in bad descriptions.

So I filed an issue: #168

And then I read the docs (as you do after first filing an issue...),
only to discover that scraping images wasn't really part of the
responsibility of metainspector [1]. Which I agree with, though the
problem for me as a consumer of the gem is that in order to really
figure out if an url serves an image, I have to first fetch it. That
feels like a lot of work, and so I started thinking about ways to make
metainspector work decently when fetching image urls. This commit is a
result of that:

 - Return `nil` for text descriptions, titles and such
 - Return the image url as the only image found

To work around an issue with grabbing content type for documents where a
`document: '<html/>'` option was passed in, I inlined the `content_type`
method instead of delegating it, and added a nil-check.

[Fixes #168]

[1]: https://github.com/jaimeiniesta/metainspector#html-content-only